requirements.txt
aws-cdk-lib>=2.136.0
constructs>=10.0.0
pydantic>=2.6.0
boto3
